fix(stackblitz), revert(dom-to-image)
