+++ packages/core/package.json
@@ -90,7 +90,6 @@
     "d3-cloud": "^1.2.5",
     "d3-sankey": "^0.12.3",
     "date-fns": "^2.30.0",
-    "dom-to-image": "^2.6.0",
     "dom-to-image-more": "^3.1.6",
     "lodash-es": "^4.17.21",
     "topojson-client": "^3.1.0",
