+++ packages/core/src/services/essentials/dom-to-image-more.d.ts
@@ -1,4 +1,4 @@
-declare module 'dom-to-image' {
+declare module 'dom-to-image-more' {
 	export type DomToImageOptions = {
 		bgcolor?: string
 		quality?: number
