+++ packages/core/vite.config.ts
@@ -46,7 +46,9 @@ export default defineConfig({
 	plugins: [
 		// equivalent to: npx tsc src/*.ts --declaration --emitDeclarationOnly --declarationDir dist/ --resolveJsonModule --esModuleInterop
 		dts({
-			exclude: './package.json'
+			exclude: './package.json',
+			skipDiagnostics: true,
+			logLevel: 'silent'
 		})
 	]
 })
